Twitter (iOS) Plugin  Issue #430
By changing the order of the if(imageAttach) and if(urlAttach) logic
the remote image preview is stacked on top of the video url reference
in the tweetViewController/Composer hiding the blank thumbnail created
by linking to a video. This creates a better UX and eliminates a
possible Apple HIG issue.